Updated span logic for when to wrap whitespace
There seems to be an exception to the rule to preserve but not wrap
whitespace: if we're directly under a P element (TODO: is it also
under a subelement? ... I need to find a kepub with this scenario),
we wrap the whitespace in a span.

This is based on looking at a few free ebooks from the last month,
like 7e01acea-41dd-4bd9-8aca-97b147a80d55 (The Great American Novel).