seriesmeta: Added support for updating SeriesID (fixes #48)
A SeriesID is required for the new series view in 4.20.14601.

This implementation is compatible with @davidfor's one for the
Calibre driver. When there is a book from the Kobo store with an
identical series name which has a series ID, that series ID is
used. Otherwise, the series ID is just the series name.

The only place where this differs from @davidfor's one is that it
doesn't care about if the author is identical. My reasoning for
this was the amount of incorrect/inconsistent (e.g. do you put a
space between multiple initials?) author information I've seen in
books, the fact that some series are written by more than one
author, and the fact that nickel handles multiple authors in a
series fine (it just merges the list of authors when it displays
the series).