Increase minimum browser versions
Even though the previous versions should technically work, and did
work with my initial prototype, they don't load the WebAssembly
code correctly (e.g. the older Chrome/Firefox versions refuse to
transfer a compiled WASM module to a worker). It's probably not
worth the hassle to try and support them since those browser
versions are a fraction of the market share, and next to none of
the actual pageviews for the kobo-related pages on my site.